[MM-27447] QA review fixes
