+++ e2e/cypress/integration/plugins/link_tooltip_spec.js
@@ -1,9 +1,15 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+// ***************************************************************
+// - [#] indicates a test step (e.g. # Go to a page)
+// - [*] indicates an assertion (e.g. * Check the title)
+// - Use element ID when selecting an element. Create one if none.
+// ***************************************************************
+
 // Group: @not_cloud @plugin
 
-import {demoPlugin} from '../../utils/plugins';
+import {demoPlugin} from '../../utils';
 
 describe('Link tooltips', () => {
     before(() => {
@@ -16,7 +22,7 @@ describe('Link tooltips', () => {
         cy.postMessage('www.test.com');
     });
 
-    it('fade in and out with an animation', () => {
+    it('MM-T3422 fade in and out with an animation', () => {
         cy.get('a[href*="www.test.com"] span').as('link');
         cy.contains('This is a custom tooltip from the Demo Plugin').parents('.tooltip-container').as('tooltip-container');
