+++ e2e/cypress/integration/search_autocomplete/scroll_spec.js
@@ -1,5 +1,12 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
+
+// ***************************************************************
+// - [#] indicates a test step (e.g. # Go to a page)
+// - [*] indicates an assertion (e.g. * Check the title)
+// - Use element ID when selecting an element. Create one if none.
+// ***************************************************************
+
 describe('Autocomplete in the search box - scrolling', () => {
     const usersCount = 15;
     const timestamp = Date.now();
@@ -14,11 +21,12 @@ describe('Autocomplete in the search box - scrolling', () => {
                 });
             }
             cy.visit(`/${team.name}/channels/town-square`);
+            cy.postMessage('hello'); // fix for https://github.com/mattermost/mattermost-webapp/pull/8084#issuecomment-844852168
         });
     });
 
     it('correctly scrolls when the user navigates through options with the keyboard', () => {
-        cy.get('#searchBox').click().type('from:');
+        cy.get('#searchBox').type('from:');
 
         cy.get('#search-autocomplete__popover .search-autocomplete__item').first().as('firstItem');
         cy.get('#search-autocomplete__popover .search-autocomplete__item').last().as('lastItem');
